+++ src/Fields/Options/Option.php
@@ -3,10 +3,13 @@
 namespace Bozboz\Jam\Fields\Options;
 
 use Bozboz\Jam\Fields\Field;
+use Bozboz\Jam\Revisionable;
 use Illuminate\Database\Eloquent\Model;
 
 class Option extends Model
 {
+    use Revisionable;
+
 	protected $table = 'entity_template_field_options';
 
 	protected $fillable = [
@@ -18,5 +21,10 @@ class Option extends Model
     public function field()
     {
         return $this->belongsTo(Field::class);
+    }
+
+    public function template()
+    {
+        return $this->field->template;
     }
 }