+++ changelog.md
@@ -1,5 +1,9 @@
 # Jam Changelog
 
+## Version 1.6.0 (Future)
+- Add seeder generator for copying template schema to production
+- Add template history tracking for easier migration of JAM schema
+
 ## Version 1.5.1 (2017-05-25)
 - Add optional format argument to published_at mutator on entity